requirements.txt
annotated-types==0.7.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
decorator==4.4.2
Flask==3.0.3
gunicorn==23.0.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
instagrapi==2.1.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
moviepy==1.0.3
numpy==2.1.2
packaging==24.2
pillow==11.0.0
proglog==0.1.10
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
setuptools==75.2.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
